requirements.txt
discord~=1.0.1
aiohttp~=3.7.4.post0
python-dotenv~=0.17.1
statsd~=3.3.0
asyncpg~=0.23.0